extract discover sections
